requirements.txt
streamlit>=1.30
pandas>=2.0
python-telegram-bot>=20.0